pyproject.toml
[project]
name = "so101-biolab-automation"
version = "0.1.0"
description = "Dual SO-101 robotic arm bio-lab automation: pipetting, tool changing, remote oversight"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "websockets>=13.0",
    "opencv-python>=4.10",
    "pyserial>=3.5",
    "pyyaml>=6.0",
    "numpy>=1.26",
]

[dependency-groups]
dev = [
    "ruff>=0.8",
    "pyright>=1.1",
]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]
train = [
    "torch>=2.4",
    "wandb>=0.18",
]

[tool.hatch.build.targets.wheel]
packages = ["src/biolab", "src/dashboard"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
default-groups = ["dev", "test"]
exclude-newer = "2026-03-25T00:00:00Z"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
include = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--strict-markers"
asyncio_mode = "auto"
markers = [
    "integration: tests loading real config files",
    "network: tests requiring network access",
    "hardware: tests requiring physical SO-101 arms or cameras",
]
